\name{unbiasedKurt}
\alias{unbiasedKurt}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal function for: Variational Bayesian Gaussian Mixture Model}
\description{
  Gives unbiased kurtosis estimate for a sample vector.
}
\usage{
unbiasedKurt(v)
}
\arguments{
  \item{v}{A vector of values, missing entries are allowed}.
}

\value{A scalar giving the kurtosis.}
  
  
\author{Andrew Teschendorff \email{aet21@hutchison-mrc.cam.ac.uk}}


\keyword{internal}% at least one, from doc/KEYWORDS

